+++ Assets/Project/Scripts/InteractionRay.cs
@@ -4,12 +4,14 @@ using UnityEngine;
 
 public class InteractionRay : MonoBehaviour
 {
-    public int length = 10;
+    public float length = 3f;
+
+    private Interactable hover;
 
 	// Use this for initialization
 	void Start ()
     {
-		
+        hover = null;
 	}
 	
 	// Update is called once per frame
@@ -17,10 +19,25 @@ public class InteractionRay : MonoBehaviour
     {
         Ray ray = new Ray(transform.position, transform.forward);
         RaycastHit hit;
-        string text;
 
-        if (Physics.Raycast(ray, out hit, length))
-            text = hit.collider.gameObject.GetComponent<Interactable>().Text;
+        Interactable objectHit = null;
 
+        if (Physics.Raycast(ray, out hit, length)) {
+            objectHit = hit.collider.gameObject.GetComponent<Interactable>();
+        }
+
+        if(objectHit != hover) {
+            if (hover != null) {
+                hover.HoverHUD.GetComponent<CanvasGroup>().alpha = 0f;
+            }
+            if (objectHit != null) {
+                objectHit.HoverHUD.GetComponent<CanvasGroup>().alpha = 1f;
+            }
+            hover = objectHit;
+        }
+
+        if(Input.GetMouseButtonDown(0) && hover != null) {
+            hover.gameObject.GetComponent<Minigame>().enter(gameObject.transform.parent.gameObject);
+        }
     }
 }